+++ example_customization_multiselect/android/app/src/main/kotlin/com/example/example_customization_multiselect/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.example.example_customization_multiselect
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}